+++ core/paddlefl_mpc/mpc_protocol/CMakeLists.txt
@@ -1,3 +1,4 @@
+aux_source_directory(./aby3_operators_impl ABY3_IMPL_SRCS)
 set(PROTO_SRCS
     "./aby3_protocol.cc"
     "./mesh_network.cc"
@@ -6,7 +7,7 @@ set(PROTO_SRCS
     "./mpc_instance.cc"
     "./mpc_protocol_factory.cc"
 )
-add_library(mpc_protocol_o OBJECT ${PROTO_SRCS})
+add_library(mpc_protocol_o OBJECT ${PROTO_SRCS} ${ABY3_IMPL_SRCS})
 add_dependencies(mpc_protocol_o fluid_framework gloo hiredis)
 
 add_library(mpc_protocol STATIC $<TARGET_OBJECTS:mpc_protocol_o>)
